+++ cfgov/cfgov/settings/local.py
@@ -2,4 +2,4 @@ from .base import *
 
 DEBUG = True
 INSTALLED_APPS += ('django_extensions',)
-STAGING_HOSTNAME = "content.localhost"
+STAGING_HOSTNAME = 'content.localhost'